requirements.txt
fastapi
uvicorn
websockets
streamlit
requests
